requirements.txt
discord.py
tabulate
requests
geopandas
pandas
purpleair
